requirements.txt
google-api-python-client>=2.70.0
google-auth>=2.20.0
google-auth-httplib2>=0.1.0
pandas>=2.0.0
openpyxl>=3.0.0
google-cloud-storage>=2.0.0
flask>=2.0.0